Refactor the PTL framework
Move all the client operations into their own component. Pull out the
common operations into base functions for reuse.
Output the test-suite.log from "make check" to help with debugging CI
failures. Break an infinite loop in pmix_tests due to incorrect return
from the event notification function.

Move the tool and server operations into their own respective components.
Pull out all the commmon operations into base functions for sharing
between the two, leaving hooks by which each can customize as required.

Try to simplify/cleanup the code using simple macros where possible to
eliminate duplicate code snippets.
