+++ .travis.yml
@@ -60,3 +60,4 @@ script:
         - make all
         # test
         - make check
+        - cat test/test-suite.log
