+++ src/include/pmix_globals.c
@@ -184,7 +184,7 @@ static void pcon(pmix_peer_t *p)
     p->proc_type.major = PMIX_MAJOR_WILDCARD;
     p->proc_type.minor = PMIX_MINOR_WILDCARD;
     p->proc_type.release = PMIX_RELEASE_WILDCARD;
-    p->proc_type.padding = 0;
+    p->proc_type.flag = 0;
     p->protocol = PMIX_PROTOCOL_UNDEF;
     p->finalized = false;
     p->info = NULL;
