+++ src/include/pmix_globals.h
@@ -164,7 +164,6 @@ typedef struct pmix_personality_t {
     pmix_bfrop_buffer_type_t type;
     pmix_bfrops_module_t *bfrops;
     pmix_psec_module_t *psec;
-    pmix_ptl_module_t *ptl;
     pmix_gds_base_module_t *gds;
 } pmix_personality_t;
 
@@ -472,7 +471,7 @@ PMIX_CLASS_DECLARATION(pmix_cb_t);
     struct timeval _tv = {0, 0};                            \
     pmix_event_evtimer_set(pmix_globals.evbase, &(r)->ev,   \
                                (c), (r));                   \
-    _tv.tv_sec = (t);                                       \
+    _tv.tv_sec = (int)(t);                                  \
     _tv.tv_usec = ((t) - _tv.tv_sec) * 1000000.0;           \
     PMIX_POST_OBJECT((r));                                  \
     pmix_event_evtimer_add(&(r)->ev, &_tv);                 \
