+++ src/mca/ptl/client/ptl_client.h
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2004-2008 The Trustees of Indiana University and Indiana
+ *                         University Research and Technology
+ *                         Corporation.  All rights reserved.
+ * Copyright (c) 2004-2006 The University of Tennessee and The University
+ *                         of Tennessee Research Foundation.  All rights
+ *                         reserved.
+ * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
+ *                         University of Stuttgart.  All rights reserved.
+ * Copyright (c) 2004-2005 The Regents of the University of California.
+ *                         All rights reserved.
+ * Copyright (c) 2016-2020 Intel, Inc.  All rights reserved.
+ * Copyright (c) 2018      IBM Corporation.  All rights reserved.
+ * $COPYRIGHT$
+ *
+ * Additional copyrights may follow
+ *
+ * $HEADER$
+ */
+
+#ifndef PMIX_PTL_CLIENT_H
+#define PMIX_PTL_CLIENT_H
+
+#include "src/mca/ptl/ptl.h"
+
+BEGIN_C_DECLS
+
+extern pmix_ptl_base_component_t mca_ptl_client_component;
+
+extern pmix_ptl_module_t pmix_ptl_client_module;
+
+END_C_DECLS
+
+#endif /* PMIX_PTL_CLIENT_H */
