+++ test/simple/Makefile.am
@@ -26,7 +26,7 @@ headers = simptest.h
 noinst_PROGRAMS = simptest simpclient simppub simpdyn simpft simpdmodex \
                   test_pmix simptool simpdie simptimeout \
                   gwtest gwclient stability quietclient simpjctrl simpio simpsched \
-                  simpcoord simpcycle doubleget simpfabric get_put_example
+                  simpcoord simpcycle doubleget simpfabric get_put_example simpvni
 
 simptest_SOURCES = \
         simptest.c
@@ -158,4 +158,10 @@ get_put_example_SOURCES = \
         get_put_example.c
 get_put_example_LDFLAGS = $(PMIX_PKG_CONFIG_LDFLAGS)
 get_put_example_LDADD = \
+    $(top_builddir)/src/libpmix.la
+
+simpvni_SOURCES = \
+        simpvni.c
+simpvni_LDFLAGS = $(PMIX_PKG_CONFIG_LDFLAGS)
+simpvni_LDADD = \
     $(top_builddir)/src/libpmix.la
